Add hook to call on result recording (#3)
When an assertion is run, Testament creates a result and adds it to a
report for the relevant test. This commit adds a hook that a user can
set to be called when the result is added to the report. This can be
used to produce streaming results, for example.

Co-authored-by: Michael Camilleri <mike@inqk.net>