Implement proprer test naming
This commit adds naming functionality to tests. Previously, it was
possible for tests with the same name to be registered in the test
suite. Results for tests were added to a test's report by assuming
that the current test was at the end of an array.

Tests now are recorded by name in the test suite and results are added
to reports by using the name of the test. Results now also include the
name of the test to which they belong as part of their data.