Support anonymous tests
While there can be benefits to naming tests, Testament's report printer
may provide enough detail in its description of the failed assertion to
identify the failing test. In such cases, the need to name a test can be
unnecessary overhead.

This commit allows a user to omit the name of the test. This works by
checking whether the first argument is a symbol. If it is, it is treated
as the name of the test. Otherwise, Testament uses a call to `gensym` to
generate a unique name that it then uses to name the test.