Add defsuite!
`defsuite` is a macro that automatically calls `run-tests!` after 
executing the body passed to it. It is intended to make (run-tests!) 
harder to forget, as I've forgotten it twice already. I prefer having something
I write up front (in the pattern of defer), rather than something I have
to remember at the end.